refactor: component "audit-result" renamed to "audit-parser-result" issue #14
