+++ projects/gpn-ui/src/app/features/audit/audit-parser-result/audit-parser-result.component.scss
